refactor: game1, 2 네이밍 수정
