+++ src/constants/game.js
@@ -1,7 +1,7 @@
-export const GAME2_ROOMS_PER_PAGE = 6;
+export const GAME1_ROOMS_PER_PAGE = 6;
 
-export const GAME2_ROUND_OPTIONS = ["1round", "2round", "3round", "4round", "5round"];
+export const GAME1_ROUND_OPTIONS = ["1round", "2round", "3round", "4round", "5round"];
 
-export const GAME2_PLAYER_OPTIONS = ["2명", "4명", "6명", "8명"];
+export const GAME1_PLAYER_OPTIONS = ["2명", "4명", "6명", "8명"];
 
-export const GAME2_LEVEL_OPTIONS = ["초보", "고수"];
+export const GAME1_LEVEL_OPTIONS = ["초보", "고수"];
