+++ src/pages/lobby/LobbyPage.jsx
@@ -27,14 +27,14 @@ const LobbyPage = () => {
           <img src={springImg} alt="스프링" className="main-content-img" />
           <div className="main-content-nav-container">
             <Link
-              to={GAME1_ROBBY_URL}
-              className={`a-btn ${navType === "game1" ? "active" : ""}`}
+              to={GAME2_ROBBY_URL}
+              className={`a-btn ${navType === "game2" ? "active" : ""}`}
             >
               루루의 미대입시
             </Link>
             <Link
-              to={GAME2_ROBBY_URL}
-              className={`b-btn ${navType === "game2" ? "active" : ""}`}
+              to={GAME1_ROBBY_URL}
+              className={`b-btn ${navType === "game1" ? "active" : ""}`}
             >
               묘묘를 속여라!
             </Link>
