+++ src/styles/pages/game/Game1CreatePage.scss
@@ -1,4 +1,4 @@
-.game2-create-container {
+.game1-create-container {
   display: flex;
   flex-direction: column;
   width: 528px;
@@ -6,7 +6,7 @@
   margin-left: 35px;
   margin-top: 17px;
 
-  .game2-create-header {
+  .game1-create-header {
     display: flex;
     justify-content: space-between;
     border-bottom: 1px solid black;
