+++ src/styles/pages/game/Game1LobbyPage.scss
@@ -1,4 +1,4 @@
-.game2-lobby-container {
+.game1-lobby-container {
   display: flex;
   flex-direction: column;
   width: 540px;
@@ -6,7 +6,7 @@
   margin-left: 35px;
   margin-top: 17px;
 
-  .game2-filter-header {
+  .game1-filter-header {
     display: flex;
     justify-content: end;
     gap: 25.28px;
